- earth and fire elemental totems work again - mob spells have a global cooldown of 2 seconds. This solves a problem when some mobs applied 3-4 negative auras on target in a very very short amount of time. - changed a few lvl 80 related values. These are pure bullshit though

+++ src/arcemu-world/BattlegroundMgr.cpp
@@ -393,7 +393,7 @@ void CBattlegroundManager::EventQueueUpdate(bool forceStart)
 					bg = iitr->second;
 					for(k = 0; k < 2; ++k)
 				{
-						int size = tempPlayerVec[k].size();
+						int size = (int)tempPlayerVec[k].size();
 						for(int counter = 0; counter < size; counter++)
 						{
 							plrguid = *tempPlayerVec[k].begin();
